feat: initial implementation
Initial implementation of the load test runner tool. Some scenarios are
unrealistic for RuloDB due to missing APIs (read-heavy, write-heavy, and
balanced), though other implemented scenarios are giving a somewhat
realistic view on the current state compared to RethinkDB.

Until the clustering and authentication are missing from RuloDB, there
are less logic slowing down the processing of queries.

Also, the RethinkDB scenario implementations should be revisited if can
be optimized (like with connection pooling if that's not automatic).

Signed-off-by: Gabor Boros <gabor.brs@gmail.com>